organize components, add save theme placeholder
