+++ src/app/colors/(playground)/page.tsx
@@ -1,10 +1,10 @@
 
-import ColorScaleGenerator from "@/components/color-scale-generator";
-export default function ColorPlayground() {
+import ColorPlayground from "@/components/color-playground";
+export default function ColorPlaygroundPage() {
   return (
     <>
       <h1 className="text-4xl font-bold">Color Playground</h1>
-      <ColorScaleGenerator />
+      <ColorPlayground />
     </>
   );
 }
