+++ src/app/colors/radix-colors/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { RADIX_REFERENCE_COLORS } from "@/CONSTANTS";
-import ReferenceColorScales from "@/components/reference-color-scale";
+import ReferenceColorScales from "@/components/color-display/reference-color-scale";
 import { useState } from "react";
 
 export default function RadixColors() {
