+++ src/app/tailwind-colors/page.tsx
@@ -6,7 +6,7 @@ import { useState } from "react";
 export default function RadixColors() {
     const [printColorSpace, setPrintColorSpace] = useState<"hsl" | "oklch">("hsl");
     return (
-      <>
+      <main className="my-12 px-4 md:px-8 lg:max-w-7xl lg:px-8 mx-auto">
         <h1 className="text-4xl font-bold mb-20">Tailwind Colors</h1>
         <button
           className="px-4 py-2 bg-slate-100 rounded-full shadow-sm text-gray-950"
@@ -18,7 +18,7 @@ export default function RadixColors() {
         </button>
 
         <ReferenceColorScales referenceColors={TAILWIND_REFERENCE_COLORS} printColorSpace={printColorSpace}/>
-      </>
+      </main>
     );
   }
