+++ src/components/color-display/color-scale.tsx
