+++ src/components/color-display/reference-color-scale.tsx
