+++ src/components/color-display/small-color-scale.tsx
