--- src/components/color-input.tsx
@@ -1,3 +1,0 @@
-export default function ColorInput() {
-  return <p>color input</p>;
-}
