+++ src/components/theme-options/save-theme.tsx
@@ -1,0 +1,7 @@
+export default function SaveTheme() {
+  return (
+    <footer>
+      <button>Save theme</button>
+    </footer>
+  );
+}
