+++ src/components/theme-options/theme-options.tsx
@@ -6,10 +6,10 @@ import { createCSSVariables } from "@/lib/theme-vars";
 import { NewColor, ReferenceColor } from "@/lib/types";
 import { useEffect, useState } from "react";
 import ButtonToggle from "./button-toggle";
-import HeadingFontSelection from "./heading-font-selection";
-import NeutralsRadio from "./neutrals-radio";
+import HeadingFontRadioGroup from "./heading-font-radio-group";
+import NeutralsRadio from "./neutrals-radio-group";
 import RadiusRadioGroup from "./radius-radio-group";
-import SmallColorScale from "./small-color-scale";
+import SmallColorScale from "../color-display/small-color-scale";
 
 export default function ThemeOptions() {
   const [brandColor, setBrandColor] = useState<string>("#a56f8e");
@@ -98,7 +98,7 @@ export default function ThemeOptions() {
       />
 
       <RadiusRadioGroup />
-      <HeadingFontSelection />
+      <HeadingFontRadioGroup />
     </div>
   );
 }
